Allow users inside container access to sudo
The crucible build requires root access inside the container. Typically, this
was enabled by just running the container as the root user, but then this
requires the user to use sudo in order to cleanup any left over artifacts.

This change will allow the container to be kicked off as a non-root user, but
still get root inside the container without `docker exec`ing from another shell.

Then the build scripts can change the owners of build artifacts before the
container is shutdown.